pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "python-docx>=1.1.2",
]
